refactor: Isolar bean do PasswordEncoder para corrigir dependencia ciclica
